# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Reservation application",
        "title": "Reservation App",
        "contact": {}
    },
    "host": "localhost:8443",
    "basePath": "/",
    "paths": {
        "/reservations": {
            "post": {
                "tags": [
                    "reservations"
                ],
                "summary": "make reservation",
                "parameters": [
                    {
                        "description": "reservation",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateReservation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/reservations/{room_id}": {
            "get": {
                "tags": [
                    "reservations"
                ],
                "summary": "Get all reservations for a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Reservation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateReservation": {
            "type": "object",
            "required": [
                "end_time",
                "room_id",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "models.Reservation": {
            "type": "object",
            "required": [
                "end_time",
                "room_id",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        }
    }
}